requirements.txt
pywin32
PyYAML
Requests
winrt
